implemented cron jobs and term state

+++ database/migrations/2024_09_22_195221_add_term_id_to_assessments.php
@@ -11,8 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
+
         Schema::table('assessments', function (Blueprint $table) {
-            $table->foreignId('term_id')->constrained('terms')->onDelete('cascade');
+            if (!Schema::hasColumn('assessments', 'term_id')) {
+                $table->foreignId('term_id')->constrained('terms')->onDelete('cascade');
+            }
         });
     }
