+++ database/migrations/2024_09_22_195923_add_term_id_to_grades.php
@@ -12,8 +12,15 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('grades', function (Blueprint $table) {
-            $table->foreignId('term_id')->constrained('terms')->onDelete('cascade');
-            $table->foreignId('class_id')->constrained('classroom');
+            // Check for the term_id column before adding it
+            if (!Schema::hasColumn('grades', 'term_id')) {
+                $table->foreignId('term_id')->constrained('terms')->onDelete('cascade');
+            }
+
+            // Check for the class_id column before adding it
+            if (!Schema::hasColumn('grades', 'class_id')) {
+                $table->foreignId('class_id')->constrained('classrooms')->onDelete('cascade');
+            }
         });
     }
 
@@ -23,7 +30,15 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('grades', function (Blueprint $table) {
-            //
+            if (Schema::hasColumn('grades', 'term_id')) {
+                $table->dropForeign(['term_id']);
+                $table->dropColumn('term_id');
+            }
+
+            if (Schema::hasColumn('grades', 'class_id')) {
+                $table->dropForeign(['class_id']);
+                $table->dropColumn('class_id');
+            }
         });
     }
 };
